+++ src/app/components/shared/card-lead/card-lead.component.ts
@@ -1,6 +1,7 @@
 import {CommonModule} from '@angular/common';
 import {ChangeDetectionStrategy, Component, Input} from '@angular/core';
 import {MatCardModule} from '@angular/material/card';
+
 import {TranslateModule} from '@ngx-translate/core';
 
 /**